@@ -1,7 +1,7 @@
 [package]
 name = "script_api"
 version = "0.0.1"
-edition = "2021"
+edition = "2024"
 description = "script_api"
 license = "GPL-3.0"
 
