@@ -1,7 +1,7 @@
 [package]
 name = "web_api"
 version = "0.0.1"
-edition = "2021"
+edition = "2024"
 description = "web_api"
 license = "GPL-3.0"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
