@@ -1,5 +1,5 @@
-use crate::diary::{service, Result};
 use crate::AppState;
+use crate::diary::{Result, service};
 use axum::extract::{Path, State};
 use axum::response::IntoResponse;
 use axum::routing::{get, post, put};
